Add noise, basic sphere ground

+++ src/image.h
@@ -6,8 +6,8 @@
 struct Image {
     uint16_t width;
     uint16_t height;
-    glm::vec3* data;
+    glm::dvec3* data;
 };
 
-void set_pixel(Image image, uint16_t x, uint16_t y, glm::vec3 color);
+void set_pixel(Image image, uint16_t x, uint16_t y, glm::dvec3 color);
 void write_image(Image image, char* path);
