+++ src/sdf.cpp
@@ -1,8 +1,8 @@
 #include "sdf.h"
 #include <cmath>
 
-SignedDistanceData sphere(glm::vec3 position) {
-    float distance = glm::length(position) - 0.5;
+SignedDistanceData sphere(glm::dvec3 position) {
+    double distance = glm::length(position) - 0.5;
     
     SignedDistanceData result;
     result.distance = distance;
@@ -10,10 +10,10 @@ SignedDistanceData sphere(glm::vec3 position) {
     return result;
 }
 
-SignedDistanceData cube(glm::vec3 position) {
-    float x_distance = fabs(position.x) - 0.5;
-    float y_distance = fabs(position.y) - 0.5;
-    float z_distance = fabs(position.z) - 0.5;
+SignedDistanceData cube(glm::dvec3 position) {
+    double x_distance = fabs(position.x) - 0.5;
+    double y_distance = fabs(position.y) - 0.5;
+    double z_distance = fabs(position.z) - 0.5;
 
     SignedDistanceData data;
     data.distance = fmax(fmax(x_distance, y_distance), z_distance);
@@ -21,9 +21,9 @@ SignedDistanceData cube(glm::vec3 position) {
     return data;
 }
 
-SignedDistanceData cylinder(glm::vec3 position) {
-    float radial_distance = sqrt(position.x * position.x + position.y * position.y) - 0.5;
-    float axial_distance = fabs(position.z) - 0.5;
+SignedDistanceData cylinder(glm::dvec3 position) {
+    double radial_distance = sqrt(position.x * position.x + position.y * position.y) - 0.5;
+    double axial_distance = fabs(position.z) - 0.5;
 
     SignedDistanceData data;
     data.distance = fmax(radial_distance, axial_distance);
@@ -31,8 +31,8 @@ SignedDistanceData cylinder(glm::vec3 position) {
     return data;
 }
 
-SignedDistanceData union_sdf(sdf_func a, sdf_func b, glm::vec3 position);
-SignedDistanceData intersect_sdf(sdf_func a, sdf_func b, glm::vec3 position);
-SignedDistanceData transform_sdf(sdf_func sdf, glm::mat4 transform, glm::vec3 position);
-SignedDistanceData fold_x(sdf_func sdf, glm::vec3 position);
-SignedDistanceData mirror_x(sdf_func sdf, glm::vec3 position);
+SignedDistanceData union_sdf(sdf_func a, sdf_func b, glm::dvec3 position);
+SignedDistanceData intersect_sdf(sdf_func a, sdf_func b, glm::dvec3 position);
+SignedDistanceData transform_sdf(sdf_func sdf, glm::dmat4 transform, glm::dvec3 position);
+SignedDistanceData fold_x(sdf_func sdf, glm::dvec3 position);
+SignedDistanceData mirror_x(sdf_func sdf, glm::dvec3 position);
